feat: melhorias e correções no projeto e-commerce

+++ src/main/java/com/github/GustavoAraujoPires/Projeto/e_commerce/service/ClienteService.java
@@ -1,8 +1,10 @@
 package com.github.GustavoAraujoPires.Projeto.e_commerce.service;
 
 import com.github.GustavoAraujoPires.Projeto.e_commerce.exception.ClienteInvalidoException;
+import com.github.GustavoAraujoPires.Projeto.e_commerce.exception.ClienteIdNaoEncontradoException;
 import com.github.GustavoAraujoPires.Projeto.e_commerce.model.Cliente;
 import com.github.GustavoAraujoPires.Projeto.e_commerce.repository.ClienteRepository;
+import com.github.GustavoAraujoPires.Projeto.e_commerce.repository.ProdutoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +14,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ClienteService {
     private final ClienteRepository repository;
+    private final ProdutoRepository produtoRepository;
 
     public Cliente salvar(Cliente cliente){
         if (repository.existsByEmail(cliente.getEmail())){
@@ -25,10 +28,17 @@ public class ClienteService {
     }
 
     public Cliente buscarPorId(Long id){
-        return repository.findById(id).orElseThrow(() -> new ClienteInvalidoException("Id não encontrado !!"));
+        return repository.findById(id).orElseThrow(() -> new ClienteIdNaoEncontradoException("Id não encontrado !!"));
     }
 
     public void deletarCliente(Long id){
-          repository.deleteById(id);
+
+       var cliente = repository.findById(id);
+        if (cliente.isPresent()){
+            repository.deleteById(id);
+        }
+        else {
+            throw new ClienteIdNaoEncontradoException("Id não encontrado !!");
+        }
     }
 }
